requirements.txt
pypdf
Pillow
PyYAML
pytest
azure-ai-documentintelligence
openai
certifi
